% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factorGettingFuncs.R
\name{gf.fin_jytable}
\alias{gf.fin_jytable}
\title{gf.fin_jytable}
\usage{
gf.fin_jytable(TS, stockID, begT, endT, type = c("PE", "PB", "PCFTTM",
  "PSTTM", "DividendRatio", "ROETTM", "EPSTTM", "NetProfitRatioTTM"))
}
\description{
get financial ratio from jy database,daily ratios such as PE,PB,DividendRatio are from table \code{LC_DIndicesForValuation},ratios from financial reports are from table \code{LC_MainIndexNew}
}
