% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factorGettingFuncs.R
\name{gf.IVR}
\alias{gf.IVR}
\title{get IVR factor}
\usage{
gf.IVR(TS, nwin = 22, datasrc = defaultDataSRC())
}
\arguments{
\item{TS}{is a TS object.}

\item{nwin}{time window.}

\item{datasrc}{is data source parameter,default value is "local".}
}
\value{
a TSF object
}
\description{
get IVR factor
}
\examples{
RebDates <- getRebDates(as.Date('2015-01-31'),as.Date('2015-12-31'),'month')
TS <- getTS(RebDates,'EI000300')
TSF <- gf.IVR(TS)
TSF <- gf.IVR(TS,datasrc = 'quant')
}
\author{
Andrew Dow
}

