% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factorGettingFuncs.R
\name{gf.val_perrank}
\alias{gf.val_perrank}
\title{gf.val_perrank}
\usage{
gf.val_perrank(TS, type = c("PE", "PB", "DividendRatio", "ROETTM", "EPSTTM",
  "NetProfitRatioTTM"), nwin = 240, cumula = FALSE, TSF)
}
\arguments{
\item{nwin}{is rolling window,default value is 240,equals one year.if \code{type} is ratios from financial reports,such as ROETTM,then nwin will become 4 inside this function.}

\item{cumula}{whether to use cumulative historical data,default value is \code{FALSE}.}

\item{TSF}{support self_defined factors as input,if TSF is not daily data,\code{nwin} need to revised.}
}
\description{
get valuation factors' percentrank,support pe,pb,roe...
}
